backend/document_backend/requirements.txt
django
python-dotenv
requests
google-genai
minio
django-cors-headers
pymysql
uuid
sentence_transformers
scikit-learn